Moved filter to Geovisto core
