+++ package.json
@@ -52,7 +52,6 @@
     "css-loader": "^5.0.1",
     "eslint": "^7.17.0",
     "font-awesome": "^4.7.0",
-    "node-sass": "^5.0.0",
     "postcss-copy": "^7.1.0",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
@@ -96,6 +95,7 @@
     "martinez-polygon-clipping": "^0.7.0",
     "osmtogeojson": "^3.0.0-beta.4",
     "polygon-clipping": "^0.15.2",
-    "rfdc": "^1.3.0"
+    "rfdc": "^1.3.0",
+    "sass": "^1.56.1"
   }
 }
