+++ src/tools/filters/index.ts
@@ -1,32 +1,26 @@
-export { GeovistoFiltersTool } from './GeovistoFiltersTool';
+export { GeovistoFiltersTool } from "./GeovistoFiltersTool";
 
-// types
-export type { default as IMapFilterManager } from './model/types/filter/IMapFilterManager';
-export type { default as IMapFilterOperation } from './model/types/filter/IMapFilterOperation';
-
-export type { default as IMapFilterRule } from './model/types/filter/IMapFilterRule';
-
-export type { default as IFiltersTool } from './model/types/tool/IFiltersTool';
-export type { default as IFiltersToolConfig } from './model/types/tool/IFiltersToolConfig';
-export type { default as IFiltersToolDefaults } from './model/types/tool/IFiltersToolDefaults';
-export type { default as IFiltersToolProps } from './model/types/tool/IFiltersToolProps';
-export type { default as IFiltersToolState } from './model/types/tool/IFiltersToolState';
+export type { default as IFiltersTool } from "./model/types/tool/IFiltersTool";
+export type { default as IFiltersToolConfig } from "./model/types/tool/IFiltersToolConfig";
+export type { default as IFiltersToolDefaults } from "./model/types/tool/IFiltersToolDefaults";
+export type { default as IFiltersToolProps } from "./model/types/tool/IFiltersToolProps";
+export type { default as IFiltersToolState } from "./model/types/tool/IFiltersToolState";
 
 // internal
-export { default as MapFilterOperation } from './model/internal/filter/basic/MapFilterOperation';
-export { default as MapFilterRule } from './model/internal/filter/basic/MapFilterRule';
-export { default as MapFiltersManager } from './model/internal/filter/basic/MapFiltersManager';
+export { default as MapFilterOperation } from "../../model/internal/filter/basic/MapFilterOperation";
+export { default as MapFilterRule } from "../../model/internal/filter/basic/MapFilterRule";
+export { default as MapFiltersManager } from "../../model/internal/filter/basic/MapFiltersManager";
 
-export { default as EqFilterOperation } from './model/internal/filter/custom/EqFilterOperation';
-export { default as GtEqFilterOperation } from './model/internal/filter/custom/GtEqFilterOperation';
-export { default as GtFilterOperation } from './model/internal/filter/custom/GtFilterOperation';
-export { default as LtEqFilterOperation } from './model/internal/filter/custom/LtEqFilterOperation';
-export { default as LtFilterOperation } from './model/internal/filter/custom/LtFilterOperation';
-export { default as NeqFilterOperation } from './model/internal/filter/custom/NeqFilterOperation';
-export { default as RegFilterOperation } from './model/internal/filter/custom/RegFilterOperation';
+export { default as EqFilterOperation } from "../../model/internal/filter/custom/EqFilterOperation";
+export { default as GtEqFilterOperation } from "../../model/internal/filter/custom/GtEqFilterOperation";
+export { default as GtFilterOperation } from "../../model/internal/filter/custom/GtFilterOperation";
+export { default as LtEqFilterOperation } from "../../model/internal/filter/custom/LtEqFilterOperation";
+export { default as LtFilterOperation } from "../../model/internal/filter/custom/LtFilterOperation";
+export { default as NeqFilterOperation } from "../../model/internal/filter/custom/NeqFilterOperation";
+export { default as RegFilterOperation } from "../../model/internal/filter/custom/RegFilterOperation";
 
-export { default as FiltersToolMapForm } from './model/internal/form/FiltersToolMapForm';
+export { default as FiltersToolMapForm } from "./model/internal/form/FiltersToolMapForm";
 
-export { default as FiltersTool } from './model/internal/tool/FiltersTool';
-export { default as FiltersToolDefaults } from './model/internal/tool/FiltersToolDefaults';
-export { default as FiltersToolState } from './model/internal/tool/FiltersToolState';
+export { default as FiltersTool } from "./model/internal/tool/FiltersTool";
+export { default as FiltersToolDefaults } from "./model/internal/tool/FiltersToolDefaults";
+export { default as FiltersToolState } from "./model/internal/tool/FiltersToolState";
