--- src/tools/filters/model/internal/filter/custom/EqFilterOperation.ts
@@ -1,39 +1,0 @@
-// Geovisto core
-import {
-    AbstractMapDomain
-} from "../../../../../../index.core";
-
-import IMapFilterOperation from "../../../types/filter/IMapFilterOperation";
-
-/**
- * This class wraps the equals filter operation.
- * 
- * @author Jiri Hynek
- */
-class EqFilterOperation extends AbstractMapDomain implements IMapFilterOperation {
-
-    /**
-     * It creates the equals filter operation.
-     */
-    public constructor() {
-        super();
-    }
-
-    /**
-     * It returns the string label of the filter representing the operator.
-     */
-    public getName(): string {
-        return "==";
-    }
-
-    /**
-     * It checks if a value equals to a pattern.
-     * 
-     * @param value 
-     * @param pattern 
-     */
-    public match(value: unknown, pattern: string): boolean {
-        return value == pattern;
-    }
-}
-export default EqFilterOperation;
